+++ RayTracerInAWeekend/material.hpp
@@ -51,16 +51,16 @@ private:
 };
 class dielectric : public material {
 public:
-    dielectric(double refraction_index) : refraction_index(refraction_index) {}
+    dielectric(float refraction_index) : refraction_index(refraction_index) {}
 
     bool scatter(const ray& r_in, const hit_record& rec, color& attenuation, ray& scattered)
         const override {
-        attenuation = color(1.0, 1.0, 1.0);
-        double ri = rec.front_face ? (1.0 / refraction_index) : refraction_index;
+        attenuation = color(1.0f, 1.0f, 1.0f);
+        float ri = rec.front_face ? (1.0f / refraction_index) : refraction_index;
 
         vec3 unit_direction = unit_vector(r_in.direction());
-        double cos_theta = fmin(dot(-unit_direction, rec.normal), 1.0);
-        double sin_theta = sqrt(1.0 - cos_theta * cos_theta);
+        float cos_theta = fmin(dot(-unit_direction, rec.normal), 1.0);
+        float sin_theta = sqrt(1.0f - cos_theta * cos_theta);
 
         bool cannot_refract = ri * sin_theta > 1.0;
         vec3 direction;
@@ -75,8 +75,8 @@ public:
     }
 
 private:
-    double refraction_index;
-    static double reflectance(double cosine, double refraction_index) {
+    float refraction_index;
+    static float reflectance(float cosine, float refraction_index) {
         auto r0 = (1 - refraction_index) / (1 + refraction_index);
         r0 = r0 * r0;
         return r0 + (1 - r0) * pow((1 - cosine), 5);
